+++ app/models/activity_item.go
@@ -10,7 +10,7 @@ import (
 type ActivityItem struct {
 	ID                   int64     `db:"id" json:"id"`
 	ActivityID           int       `db:"activity_id" json:"activity_id"`
-	EventID              int       `db:"event_id" json:"event_id"`
+	EventID              int64     `db:"event_id" json:"event_id"`
 	Title                string    `db:"title" json:"title"`
 	Memo                 string    `db:"memo" json:"memo"`
 	Notification         bool      `db:"notification" json:"notification"`
